Remove structs.h
- Remove structs.h
- Distribute some of the structure to where they are used, eliminate other uses
- Remove some unused code
- Slightly reduce the dependencies animation/rendering related code has on ChkdraftLib-level files
